moved url field to file
